+++ pkg/chunk/cache/snappy.go
@@ -2,6 +2,7 @@ package cache
 
 import (
 	"context"
+	"time"
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
@@ -23,17 +24,17 @@ func NewSnappy(next Cache, logger log.Logger) Cache {
 	}
 }
 
-func (s *snappyCache) Store(ctx context.Context, keys []string, bufs [][]byte) {
+func (s *snappyCache) Store(ctx context.Context, keys []string, bufs [][]byte, ttl time.Duration) {
 	cs := make([][]byte, 0, len(bufs))
 	for _, buf := range bufs {
 		c := snappy.Encode(nil, buf)
 		cs = append(cs, c)
 	}
-	s.next.Store(ctx, keys, cs)
+	s.next.Store(ctx, keys, cs, ttl)
 }
 
-func (s *snappyCache) Fetch(ctx context.Context, keys []string) ([]string, [][]byte, []string) {
-	found, bufs, missing := s.next.Fetch(ctx, keys)
+func (s *snappyCache) Fetch(ctx context.Context, keys []string, ttl time.Duration) ([]string, [][]byte, []string) {
+	found, bufs, missing := s.next.Fetch(ctx, keys, ttl)
 	ds := make([][]byte, 0, len(bufs))
 	for _, buf := range bufs {
 		d, err := snappy.Decode(nil, buf)